phase1/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
databases
pydantic
pdfminer.six
chromadb
boto3
python-dotenv
requests